requirements.txt
asgiref==3.8.1
Django==5.2.1
django-ckeditor==6.7.2
django-js-asset==3.1.2
fastjsonschema==2.21.1
pillow==11.3.0
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.2
python-json-logger==3.3.0
PyYAML==6.0.2
pyzmq==27.0.0
rfc3986-validator==0.1.1
rpds-py==0.26.0
Send2Trash==1.8.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
tinycss2==1.4.0
tornado==6.5.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250708
typing_extensions==4.14.1
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
